+++ internal/token/token.go
@@ -1,4 +1,4 @@
-package lexer
+package token
 
 type TokenType int
